+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2024-04-29-1
+
+Changes:
+
+- Seperate system language checks
+- Rename `fallback` properties
+
 ## v2024-04-22-2
 
 Changes: